Pulled txStatusOption out from handlers\package.scala, and moved it into HandlerImpls
