+++ node/src/main/scala/co/topl/rpc/handlers/package.scala
@@ -5,7 +5,6 @@ import co.topl.attestation.Address
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.Block
 import co.topl.network.message.BifrostSyncInfo
-import co.topl.nodeView.ReadableNodeView
 import co.topl.nodeView.history.HistoryReader
 import co.topl.nodeView.state.StateReader
 import co.topl.rpc.ToplRpc.NodeView.ConfirmationStatus.TxStatus
@@ -65,23 +64,8 @@ package object handlers {
     } yield (startHeight, endHeight)
 
   private[handlers] def checkTxIds(
-    txIds: List[ModifierId],
-    view:  ReadableNodeView
-  ): Either[RpcError, ToplRpc.NodeView.ConfirmationStatus.Response] = {
-    val txStatusOption: List[Option[(ModifierId, TxStatus)]] = txIds.map { id =>
-      val mempoolStatus = view.memPool.modifierById(id)
-      val historyStatus = view.history.transactionById(id)
-      val bestBlockHeight = view.history.height
-      (mempoolStatus, historyStatus) match {
-        case (_, Some((tx, _, height))) =>
-          Some(tx.id -> TxStatus("Confirmed", bestBlockHeight - height))
-        case (Some(tx), None) =>
-          Some(tx.id -> TxStatus("Unconfirmed", -1))
-        case (None, None) =>
-          None
-      }
-    }
-
+    txStatusOption: List[Option[(ModifierId, TxStatus)]]
+  ): Either[RpcError, ToplRpc.NodeView.ConfirmationStatus.Response] =
     for {
       txStatus <- Either.cond(
         txStatusOption.forall(_.nonEmpty),
@@ -89,6 +73,5 @@ package object handlers {
         ToplRpcErrors.NoTransactionWithId
       )
     } yield txStatus
-  }
 
 }
